Improves asignatura query by campoFormacion
Updates the query for retrieving asignaturas by campoFormacion to handle the case where campoFormacion is 'Otros'.
When campoFormacion is 'Otros', it now retrieves asignaturas where campo_formacion is an empty string.
For other values of campoFormacion, it uses a LIKE clause to find asignaturas where campo_formacion contains the specified campoFormacion value.